% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publicschools2016secondary.R
\docType{data}
\name{publicschools2016secondary}
\alias{publicschools2016secondary}
\title{publicschools2016secondary: a 2015-2016 secondary schools sf point file for Hamilton, Ontario.}
\format{Simple feature collection with 13 features and 8 fields:
\describe{
  \item{SCHNAME}{Name of school}
  \item{X}{the school's X co-ordinate in degrees}
  \item{Y}{the school's Y co-ordinate in degrees}
  \item{Address}{the school's Street address}
  \item{City}{the school's City}
  \item{PostalCode}{the school's postal code}
  \item{Year_Built}{the year the school was built}
  \item{SchoolID}{an ID number used by the Hamilton Board of Education, one unique number per school}
  \item{geometry}{the simple feature geometry}
}}
\source{
Hamilton Board of Education
}
\usage{
data(publicschools2016secondary)
}
\description{
This dataset contains an sf points file in NAD83 UTM Zone 17N, for each secondary school
in the Hamilton public board.
}
\keyword{Hamilton}
\keyword{datasets,}
\keyword{schools,}
