% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sess7r.R
\docType{package}
\name{sess7r}
\alias{sess7r}
\alias{sess7r-package}
\title{sess7r: A package with a minimum example of package creation.}
\description{
This package is an exercise in package creation using
R studio. The package will hopefully include a sample
function and a sample dataset with their respective
documentation, though the last time I tried this it
screwed up miserably.
}
\references{
\url{https://github.com/merralja/SchoolProbit}
}
\author{
John Merrall, McMaster University \email{merralja@mcmaster.ca}
}
